+++ site/src/components/varbox/inputs/NumberInput.vue
