+++ site/src/test/varbox/VarInputCases.ts
@@ -4,7 +4,7 @@ import { createTreeFromConfig, createNodeFromConfig, cns } from "../../utils/Var
 
 // TestCase结构：键为描述，值为NodeTree对象
 export type TestCase = Record<string, { tree: VarTree}>;
-import MyCustomInput from '@/test/VarBox/MyCustomInput.vue'
+import MyCustomInput from '@/test/varbox/MyCustomInput.vue'
 
 function dummyComponent(): Component {
   // 伪造一个外部组件钩子
